Second experiments - relaxed data types and ignored some abnormal behaviors

+++ src/data_types.hpp
@@ -21,8 +21,8 @@ typedef uint16_t SequenceSize;
 typedef uint16_t StratumSize; 
 /* Can covering region be over 255? meaning stratified regions are at least 122? */ 
 /*I doubt about it even if the dataset is very largein normal cases */
-typedef uint8_t CoveringRegionIdx; 
-typedef uint8_t CoveringRegionCount; 
+typedef uint32_t CoveringRegionIdx; 
+typedef uint32_t CoveringRegionCount; 
 //Alphabet
 typedef uint8_t CharId; 
 typedef uint8_t CharCount;
